Reorganize bitset code.
Move BitSet and BitPage into separate files.
